solidago: define PipelineParameters base class
