+++ solidago/src/solidago/pipeline/__init__.py
@@ -1,1 +1,2 @@
 from .inputs import TournesolInput
+from .parameters import PipelineParameters
