+++ solidago/src/solidago/pipeline/inputs.py
@@ -8,7 +8,6 @@ import pandas as pd
 
 
 class TournesolInput(ABC):
-    SCALING_CALIBRATION_MIN_ENTITIES_TO_COMPARE = 20
     SCALING_CALIBRATION_MIN_TRUST_SCORE = 0.1
     MAX_SCALING_CALIBRATION_USERS = 100
